[depthFirstSearch] Created traversal directory & moved files in it
