+++ configuration.conf
@@ -40,7 +40,7 @@ version             | Y | Y | Y
 topologicalSort     | Y | Y | Y
 transitiveClosure   | Y | Y | Y
 breadthFirstSearch  | Y | Y | Y
-depthFirstSearch    | Y | Y | Y
+traversal           | Y | Y | Y
 coloring            | Y | Y | Y
 #----------------------
 # SQL only directories
