+++ doc/traversal/images/CMakeLists.txt
