+++ doc/traversal/images/Fig1-AscendingOrder.png
