+++ doc/traversal/images/Fig2-DescendingOrder.png
