+++ pgtap/traversal/edge_cases_directed.sql
