+++ pgtap/traversal/edge_cases_undirected.sql
