+++ pgtap/traversal/equivalence_BFS.test.sql
