+++ pgtap/traversal/inner_query.sql
