+++ pgtap/traversal/issue1348.sql
