+++ pgtap/traversal/no_crash_test.sql
