+++ pgtap/traversal/singleVertex_equiv_multipleVertices.test.sql
