+++ pgtap/traversal/types_check.sql
