+++ sql/traversal/_depthFirstSearch.sql
