+++ src/traversal/depthFirstSearch.c
@@ -42,7 +42,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_common/edges_input.h"
 #include "c_common/arrays_input.h"
 
-#include "drivers/depthFirstSearch/depthFirstSearch_driver.h"
+#include "drivers/traversal/depthFirstSearch_driver.h"
 
 PGDLLEXPORT Datum _pgr_depthfirstsearch(PG_FUNCTION_ARGS);
 PG_FUNCTION_INFO_V1(_pgr_depthfirstsearch);
